+++ CHANGELOG.md
@@ -3,7 +3,15 @@
 All notable changes to this project will be documented in this file.
 
 <!-- git-cliff-unreleased-start -->
-## 0.6.4 - **not yet released**
+## 0.7.1 - **not yet released**
+
+### 🚀 Features
+
+- Add default 1 hour test run timeout ([#79](https://github.com/apify/apify-test-tools/pull/79)) ([c438350](https://github.com/apify/apify-test-tools/commit/c4383509b49228e251594b4b0a5c48fde94e71f7)) by [@Patai5](https://github.com/Patai5)
+
+
+<!-- git-cliff-unreleased-end -->
+## [0.7.0](https://github.com/apify/apify-test-tools/releases/tag/v0.7.0) (2026-04-10)
 
 ### 🚀 Features
 
@@ -13,8 +21,11 @@ All notable changes to this project will be documented in this file.
 
 - Standby tests ([#76](https://github.com/apify/apify-test-tools/pull/76)) ([33cd89c](https://github.com/apify/apify-test-tools/commit/33cd89cf9e512f422704eeb8f6c46e4e2759a902)) by [@oklinov](https://github.com/oklinov)
 
+### ⚡ Performance
 
-<!-- git-cliff-unreleased-end -->
+- **bin:** Don&#x27;t test if we merge master into cosmetic only changes ([#78](https://github.com/apify/apify-test-tools/pull/78)) ([bc899eb](https://github.com/apify/apify-test-tools/commit/bc899eb550283862fbdc3c44a4fca6f682017c43)) by [@metalwarrior665](https://github.com/metalwarrior665)
+
+
 ## [0.6.3](https://github.com/apify/apify-test-tools/releases/tag/v0.6.3) (2026-04-06)
 
 ### 🐛 Bug Fixes